+++ gl_program.c
@@ -399,12 +399,19 @@ static void GL_ConPrintShaderLog(GLuint shader)
 
 	GL_Procedure(glGetShaderiv, shader, GL_INFO_LOG_LENGTH, &log_length);
 	GL_Procedure(glGetShaderiv, shader, GL_SHADER_SOURCE_LENGTH, &src_length);
-	if (log_length) {
+	if (log_length || (src_length > 0 && developer.integer == 3)) {
 		GLsizei written;
 
-		buffer = Q_malloc(max(log_length, src_length));
+		buffer = Q_malloc(max(log_length, src_length) + 1);
 		GL_Procedure(glGetShaderInfoLog, shader, log_length, &written, buffer);
-		Con_Printf(buffer);
+		Con_Printf("-- Version: %s\n", glConfig.version_string);
+		Con_Printf("--    GLSL: %s\n", glConfig.glsl_version);
+		Con_Printf("%s\n", buffer);
+
+		if (developer.integer == 2 || developer.integer == 3) {
+			GL_Procedure(glGetShaderSource, shader, src_length, &written, buffer);
+			Con_Printf("%s\n", buffer);
+		}
 
 		Q_free(buffer);
 	}
@@ -466,6 +473,10 @@ static qbool GL_CompileShader(GLsizei shaderComponents, const char* shaderText[]
 		GL_Procedure(glGetShaderiv, shader, GL_COMPILE_STATUS, &result);
 		if (result) {
 			*shaderId = shader;
+			if (developer.integer == 3) {
+				Con_Printf("Shader->Compile(%X) succeeded\n", shaderType);
+				GL_ConPrintShaderLog(shader);
+			}
 			return true;
 		}
 
@@ -569,7 +580,6 @@ static int GL_InsertDefinitions(
 		strings[2] = (const char*)glsl_common_glsl;
 		strings[1] = (const char*)glsl_constants_glsl;
 
-		return 6;
 		// Some drivers interpret length 0 as nul terminated
 		// spec is < 0 (https://www.khronos.org/registry/OpenGL-Refpages/gl4/html/glShaderSource.xhtml)
 		for (i = 0; i < MAX_SHADER_COMPONENTS; ++i) {
@@ -578,6 +588,7 @@ static int GL_InsertDefinitions(
 			}
 		}
 
+		return 6;
 	}
 
 	return 1;